requirements.txt
streamlit==1.28.1
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.1
plotly==5.15.0
numpy==1.24.3
joblib==1.3.2
protobuf==4.24.0
